🐛 Fix : 위젯 추가 모달 내부 컴포넌트 분리, 내부에서 state 관리
