+++ src/components/home/MyWidgets.tsx
@@ -3,9 +3,17 @@ import { WidgetProps } from "@/constants/widgets";
 import H from "@/components/home";
 import { useAtom } from "jotai";
 import { myWidgets } from "@/stores/atoms";
+import { useEffect } from "react";
 
 const MyWidgets = () => {
-  const [widgets] = useAtom(myWidgets);
+  const [widgets, setWidgets] = useAtom(myWidgets);
+  useEffect(() => {
+    if (widgets.length === 0) {
+      const storedData = localStorage.getItem("selectedWidgets");
+      const widgets = storedData ? JSON.parse(storedData) : [];
+      setWidgets(widgets);
+    }
+  });
 
   return widgets.map((item: WidgetProps) => (
     <H.Widget
