+++ src/stores/atoms.ts
@@ -1,7 +1,6 @@
+import { WidgetProps } from "@/constants/widgets";
 import { atom } from "jotai";
 
 //Widget 전역상태
-const storedData = localStorage.getItem("selectedWidgets");
-const widgets = storedData ? JSON.parse(storedData) : [];
 
-export const myWidgets = atom(widgets);
+export const myWidgets = atom<WidgetProps[]>([]);
